requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.3.1
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.2.1
django-filter==25.1
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-types==0.20.0
djangorestframework==3.16.0
djangorestframework-stubs==3.16.0
filelock==3.18.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
identify==2.6.12
idna==3.10
isort==6.0.1
mypy==1.16.0
mypy_extensions==1.1.0
nodeenv==1.9.1
pathspec==0.12.1
platformdirs==4.3.8
pre-commit==4.2.0
psutil==7.0.0
psycopg==3.2.9
pycparser==2.22
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.12
setuptools==80.8.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
types-psycopg2==2.9.21.20250516
types-PyYAML==6.0.12.20250516
types-requests==2.32.0.20250515
typing_extensions==4.13.2
urllib3==2.4.0
virtualenv==20.31.2
wheel==0.45.1